bugfix : like function

+++ src/Controleur/ControleurLike.php
@@ -2,6 +2,7 @@
 
 namespace App\Altius\Controleur;
 
+use App\Altius\Lib\ConnexionUtilisateur;
 use App\Altius\Modele\DataObject\Like;
 use App\Altius\Modele\Repository\LikeRepository;
 
@@ -9,16 +10,18 @@ class ControleurLike
 {
 
     public static function like() {
-        //TODO : userID should be connected user
-        $userID = "test";
-        $like = new Like((int)$_REQUEST["publicationID"], $userID);
-        (new LikeRepository())->create($like);
+        $userID = ConnexionUtilisateur::getLoginUtilisateurConnecte();
+        if($userID != null) {
+            $like = new Like((int)$_REQUEST["publicationID"], $userID);
+            (new LikeRepository())->create($like);
+        }
     }
 
     public static function unlike() {
-        //TODO: userID should be connected user
-        $userID = "test";
-        (new LikeRepository())->deleteByID(array ($_REQUEST["publicationID"], $userID));
+        $userID = ConnexionUtilisateur::getLoginUtilisateurConnecte();
+        if ($userID!=null) {
+            (new LikeRepository())->deleteByID(array ($_REQUEST["publicationID"], $userID));
+        }
     }
